Reorganization of query runners to propagate exceptions.
